good api for new method

+++ arduino/sketchbook/libraries/sim800/sim800Client.h
@@ -18,7 +18,7 @@ class sim800Client : public SIM800 {
   uint8_t connected();
   int available();
   int read();
-  int readBytes(char *buf, size_t size);
+  byte readBytes(char *buf, size_t size);
   void setTimeout(unsigned long timeout);
   size_t write(uint8_t);
   size_t write(const uint8_t *buf, size_t size);
